Refactor TypeScript ingredient generation to support JavaScript generation
