+++ recipe-generator/src/main/java/ca/derekcormier/recipe/generator/CookbookGeneratorFactory.java
@@ -9,6 +9,8 @@ public class CookbookGeneratorFactory {
                 return new JavaHookGenerator(cookbook);
             case JAVA_INGREDIENT:
                 return new JavaIngredientGenerator(cookbook);
+            case JAVASCRIPT_INGREDIENT:
+                return new JavaScriptIngredientGenerator(cookbook);
             case TYPESCRIPT_INGREDIENT:
                 return new TypeScriptIngredientGenerator(cookbook);
             default:
