+++ recipe-generator/src/main/java/ca/derekcormier/recipe/generator/Flavour.java
@@ -3,6 +3,7 @@ package ca.derekcormier.recipe.generator;
 public enum Flavour {
     JAVA_HOOK("java-hook"),
     JAVA_INGREDIENT("java-ingredient"),
+    JAVASCRIPT_INGREDIENT("js-ingredient"),
     TYPESCRIPT_INGREDIENT("ts-ingredient");
 
     private String alias;
