+++ recipe-generator/src/main/resources/templates/js/enum.liquid
