+++ recipe-generator/src/main/resources/templates/js/ingredient-index.liquid
