+++ recipe-generator/src/test/java/ca/derekcormier/recipe/generator/filter/TsValueFilterTest.java
@@ -16,7 +16,7 @@ public class TsValueFilterTest {
     @Test(expected = IllegalArgumentException.class)
     public void testApply_throwsOnMissingTypeParam() {
         Cookbook cookbook = new Cookbook(new ArrayList<>(), new ArrayList<>());
-        Filter filter = new TsValueFilter(cookbook);
+        Filter filter = new JsValueFilter(cookbook);
 
         filter.apply("foobar");
     }
@@ -24,7 +24,7 @@ public class TsValueFilterTest {
     @Test
     public void testApply_convertsString() {
         Cookbook cookbook = new Cookbook(new ArrayList<>(), new ArrayList<>());
-        Filter filter = new TsValueFilter(cookbook);
+        Filter filter = new JsValueFilter(cookbook);
 
         assertEquals("\"foobar\"", filter.apply("foobar", "string"));
     }
@@ -32,7 +32,7 @@ public class TsValueFilterTest {
     @Test
     public void testApply_convertsStringWithQuotes() {
         Cookbook cookbook = new Cookbook(new ArrayList<>(), new ArrayList<>());
-        Filter filter = new TsValueFilter(cookbook);
+        Filter filter = new JsValueFilter(cookbook);
 
         assertEquals("\"\\\"foobar\"", filter.apply("\"foobar", "string"));
     }
@@ -40,7 +40,7 @@ public class TsValueFilterTest {
     @Test
     public void testApply_convertsNullString() {
         Cookbook cookbook = new Cookbook(new ArrayList<>(), new ArrayList<>());
-        Filter filter = new TsValueFilter(cookbook);
+        Filter filter = new JsValueFilter(cookbook);
 
         assertEquals("null", filter.apply(null, "string"));
     }
@@ -48,7 +48,7 @@ public class TsValueFilterTest {
     @Test
     public void testApply_convertsInt() {
         Cookbook cookbook = new Cookbook(new ArrayList<>(), new ArrayList<>());
-        Filter filter = new TsValueFilter(cookbook);
+        Filter filter = new JsValueFilter(cookbook);
 
         assertEquals("5", filter.apply(new Integer(5), "int"));
         assertEquals("5", filter.apply(5, "int"));
@@ -57,7 +57,7 @@ public class TsValueFilterTest {
     @Test
     public void testApply_convertsNegativeInt() {
         Cookbook cookbook = new Cookbook(new ArrayList<>(), new ArrayList<>());
-        Filter filter = new TsValueFilter(cookbook);
+        Filter filter = new JsValueFilter(cookbook);
 
         assertEquals("-1", filter.apply(new Integer(-1), "int"));
         assertEquals("-1", filter.apply(-1, "int"));
@@ -66,7 +66,7 @@ public class TsValueFilterTest {
     @Test
     public void testApply_convertsFloat() {
         Cookbook cookbook = new Cookbook(new ArrayList<>(), new ArrayList<>());
-        Filter filter = new TsValueFilter(cookbook);
+        Filter filter = new JsValueFilter(cookbook);
 
         assertEquals("5.0", filter.apply(new Float(5.0), "float"));
         assertEquals("5.0", filter.apply(5.0f, "float"));
@@ -75,7 +75,7 @@ public class TsValueFilterTest {
     @Test
     public void testApply_convertsNegativeFloat() {
         Cookbook cookbook = new Cookbook(new ArrayList<>(), new ArrayList<>());
-        Filter filter = new TsValueFilter(cookbook);
+        Filter filter = new JsValueFilter(cookbook);
 
         assertEquals("-5.0", filter.apply(new Float(-5.0), "float"));
         assertEquals("-5.0", filter.apply(-5.0f, "float"));
@@ -84,7 +84,7 @@ public class TsValueFilterTest {
     @Test
     public void testApply_convertsBoolean() {
         Cookbook cookbook = new Cookbook(new ArrayList<>(), new ArrayList<>());
-        Filter filter = new TsValueFilter(cookbook);
+        Filter filter = new JsValueFilter(cookbook);
 
         assertEquals("true", filter.apply(new Boolean(true), "boolean"));
         assertEquals("false", filter.apply(false, "boolean"));
@@ -93,7 +93,7 @@ public class TsValueFilterTest {
     @Test
     public void testApply_convertsFlag() {
         Cookbook cookbook = new Cookbook(new ArrayList<>(), new ArrayList<>());
-        Filter filter = new TsValueFilter(cookbook);
+        Filter filter = new JsValueFilter(cookbook);
 
         assertEquals("true", filter.apply(new Boolean(true), "flag"));
         assertEquals("false", filter.apply(false, "flag"));
@@ -102,7 +102,7 @@ public class TsValueFilterTest {
     @Test
     public void testApply_convertsEnum() {
         Cookbook cookbook = new Cookbook(new ArrayList<>(), Lists.newArrayList(new Enum("MyEnum", Lists.newArrayList("A"))));
-        Filter filter = new TsValueFilter(cookbook);
+        Filter filter = new JsValueFilter(cookbook);
 
         assertEquals("MyEnum.A", filter.apply("A", "MyEnum"));
     }
@@ -110,7 +110,7 @@ public class TsValueFilterTest {
     @Test
     public void testApply_convertsArrayOfStrings() {
         Cookbook cookbook = new Cookbook(new ArrayList<>(), new ArrayList<>());
-        Filter filter = new TsValueFilter(cookbook);
+        Filter filter = new JsValueFilter(cookbook);
 
         assertEquals("[\"a\", \"b\", \"c\"]", filter.apply(Lists.newArrayList("a", "b", "c"), "string[]"));
     }
@@ -118,7 +118,7 @@ public class TsValueFilterTest {
     @Test
     public void testApply_convertsArrayOfInts() {
         Cookbook cookbook = new Cookbook(new ArrayList<>(), new ArrayList<>());
-        Filter filter = new TsValueFilter(cookbook);
+        Filter filter = new JsValueFilter(cookbook);
 
         assertEquals("[1, 2, 3]", filter.apply(Lists.newArrayList(1, 2, 3), "int[]"));
     }
@@ -126,7 +126,7 @@ public class TsValueFilterTest {
     @Test
     public void testApply_convertsArrayOfEnums() {
         Cookbook cookbook = new Cookbook(new ArrayList<>(), Lists.newArrayList(new Enum("MyEnum", Lists.newArrayList("A", "B", "C"))));
-        Filter filter = new TsValueFilter(cookbook);
+        Filter filter = new JsValueFilter(cookbook);
 
         assertEquals("[MyEnum.A, MyEnum.B, MyEnum.C]", filter.apply(Lists.newArrayList("A", "B", "C"), "MyEnum[]"));
     }
@@ -134,7 +134,7 @@ public class TsValueFilterTest {
     @Test
     public void testApply_convertsVarargOfStrings() {
         Cookbook cookbook = new Cookbook(new ArrayList<>(), new ArrayList<>());
-        Filter filter = new TsValueFilter(cookbook);
+        Filter filter = new JsValueFilter(cookbook);
 
         assertEquals("[\"a\", \"b\", \"c\"]", filter.apply(Lists.newArrayList("a", "b", "c"), "string..."));
     }
@@ -142,7 +142,7 @@ public class TsValueFilterTest {
     @Test
     public void testApply_convertsVarargOfArraysOfStrings() {
         Cookbook cookbook = new Cookbook(new ArrayList<>(), new ArrayList<>());
-        Filter filter = new TsValueFilter(cookbook);
+        Filter filter = new JsValueFilter(cookbook);
 
         assertEquals("[[\"a\"], [\"b\", \"c\"]]", filter.apply(Lists.newArrayList(Lists.newArrayList("a"), Lists.newArrayList("b", "c")), "string[]..."));
     }
